Diffusion tensor imaging data

Fractional anisotropy (FA) tract profiles for the corpus callosum (cca) and the right corticospinal tract (rcst). Accompanying the tract profiles are the subject ID numbers, visit number, total number of scans, multiple sclerosis case status and Paced Auditory Serial Addition Test (pasat) score.

Data are also include in the refund package in another format.

Usage

dti

Format

A tibble with 382 rows and 6 variables:

id

Subject ID

visit

The visit number

sex

The sex of the subject

case

The multiple sclerosis case status

cca

The fractional anisotropy (FA) tract profile for the corpus callosum

rcst

The fractional anisotropy (FA) tract profile for the right corticospinal tract

Details

If you use this data as an example in written work, please include the following acknowledgment: “The MRI/DTI data were collected at Johns Hopkins University and the Kennedy-Krieger Institute"

Examples

dti
#> # A tibble: 382 × 6
#>       id visit sex    case                                                  cca
#>    <dbl> <int> <fct>  <fct>                                          <tfd_irrg>
#>  1  1001     1 female control  [1]: (0.000,0.49);(0.011,0.52);(0.022,0.54); ...
#>  2  1002     1 female control  [2]: (0.000,0.47);(0.011,0.49);(0.022,0.50); ...
#>  3  1003     1 male   control  [3]: (0.000,0.50);(0.011,0.51);(0.022,0.54); ...
#>  4  1004     1 male   control  [4]: (0.000,0.40);(0.011,0.42);(0.022,0.44); ...
#>  5  1005     1 male   control  [5]: (0.000,0.40);(0.011,0.41);(0.022,0.40); ...
#>  6  1006     1 male   control  [6]: (0.000,0.45);(0.011,0.45);(0.022,0.46); ...
#>  7  1007     1 male   control  [7]: (0.000,0.55);(0.011,0.56);(0.022,0.56); ...
#>  8  1008     1 male   control  [8]: (0.000,0.45);(0.011,0.48);(0.022,0.50); ...
#>  9  1009     1 male   control  [9]: (0.000,0.50);(0.011,0.51);(0.022,0.52); ...
#> 10  1010     1 male   control [10]: (0.000,0.46);(0.011,0.47);(0.022,0.48); ...
#> # ℹ 372 more rows
#> # ℹ 1 more variable: rcst <tfd_irrg>